+++ lib/birdiy_web/views/pow/session_view.ex
@@ -1,0 +1,7 @@
+defmodule BirdiyWeb.Pow.SessionView do
+  use BirdiyWeb, :view
+
+  def title("new.html", _assigns) do
+    "Login to Birdiy Admin"
+  end
+end