+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Admin panel.
+- Pow as admin authentication.
 
 ## [0.0.1] - 2019-08-03
 ### Added